# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: K&S Solar Energy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/checkin:
    post:
      tags: [attendance]
      operationId: checkIn
      summary: Technician check-in with GPS and photos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '200':
          description: Checked in successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
  /attendance/{id}/checkout:
    post:
      tags: [attendance]
      operationId: checkOut
      summary: Technician check-out
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Checked out successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
  /attendance:
    get:
      tags: [attendance]
      operationId: getAttendance
      summary: Get all attendance records (admin)
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: technicianId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of attendance records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceRecord'
...
  /attendance/absent-today:
    get:
      tags: [attendance]
      operationId: getAbsentToday
      summary: Get technicians who have not checked in today (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of absent technicians today.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbsentTechnician'
  /technician-locations:
    post:
      tags: [attendance]
      operationId: upsertTechnicianLocation
      summary: Upsert technician's latest location (technician only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertTechnicianLocationInput'
      responses:
        '200':
          description: Location upserted successfully.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        '400':
          description: Bad Request
        '403':
          description: Forbidden
    get:
      tags: [attendance]
      operationId: getTechnicianLiveLocations
      summary: Get latest live location for all currently checked-in technicians (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest location ping per checked-in technician.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TechnicianLiveLocation'
        '403':
          description: Forbidden
  /technician-locations/{technicianId}/trail:
    get:
      tags: [attendance]
      operationId: getTechnicianTrail
      summary: Get rolling last-24h GPS trail for a technician across sessions (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: technicianId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ordered list of GPS pings for the technician in the last 24 hours.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationPing'
        '403':
          description: Forbidden
...
  /attendance/{id}/location-trail:
    get:
      tags: [attendance]
      operationId: getLocationTrail
      summary: Get GPS location trail for an attendance session (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ordered list of GPS pings for this attendance session.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationPing'
  /attendance/{id}:
    patch:
      tags: [attendance]
      operationId: adminUpdateAttendance
      summary: Admin edit or correct a technician's check-in or check-out time
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateAttendanceRequest'
      responses:
        '200':
          description: Updated attendance record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '404':
          description: Record not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /attendance/manual:
    post:
      tags: [attendance]
      operationId: adminAddManualAttendance
      summary: Admin manually add an attendance entry for any technician
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualAttendanceRequest'
      responses:
        '201':
          description: Manually created attendance record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CheckInRequest:
      type: object
      properties:
        selfieUrl:
          type: string
          nullable: true
        sitePhotoUrl:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        locationAddress:
          type: string
          nullable: true
        siteId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    AttendanceRecord:
      type: object
      required: [id, technicianId, technicianName, checkInAt, createdAt]
      properties:
        id:
          type: string
        technicianId:
          type: string
        technicianName:
          type: string
        siteId:
          type: string
          nullable: true
        siteName:
          type: string
          nullable: true
        selfieUrl:
          type: string
          nullable: true
        sitePhotoUrl:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        locationAddress:
          type: string
          nullable: true
        checkInAt:
          type: string
        checkOutAt:
          type: string
          nullable: true
        totalHours:
          type: number
          nullable: true
        overtimeHours:
          type: number
          nullable: true
        isLate:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
    AdminUpdateAttendanceRequest:
      type: object
      properties:
        checkInAt:
          type: string
          description: ISO datetime string for the corrected check-in time
        checkOutAt:
          type: string
          nullable: true
          description: ISO datetime string for the corrected check-out time, or null to clear it
        notes:
          type: string
          nullable: true
    ManualAttendanceRequest:
      type: object
      required: [technicianId, checkInAt]
      properties:
        technicianId:
          type: string
        checkInAt:
          type: string
          description: ISO datetime string for the check-in time
        checkOutAt:
          type: string
          nullable: true
          description: ISO datetime string for the check-out time
        notes:
          type: string
          nullable: true
...
    AbsentTechnician:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
...
    LocationPing:
      type: object
      required: [id, attendanceId, technicianId, latitude, longitude, recordedAt]
      properties:
        id:
          type: string
        attendanceId:
          type: string
        technicianId:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        address:
          type: string
          nullable: true
        recordedAt:
          type: string
    UpsertTechnicianLocationInput:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: string
        longitude:
          type: string
        address:
          type: string
          nullable: true
    TechnicianLiveLocation:
      type: object
      required: [technicianId, name, latitude, longitude, recordedAt, status, isStale]
      properties:
        technicianId:
          type: string
        name:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        address:
          type: string
          nullable: true
        recordedAt:
          type: string
        attendanceId:
          type: string
          nullable: true
        checkInAt:
          type: string
          nullable: true
        status:
          type: string
          description: "Technician field status: checked-in | off-shift | offline"
        isStale:
          type: boolean
...