# lib/api-zod/src/generated/types/technicianLiveLocation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * K&S Solar Energy API
 * OpenAPI spec version: 0.1.0
 */

export interface TechnicianLiveLocation {
  technicianId: string;
  name: string;
  latitude: string;
  longitude: string;
  address?: string | null;
  recordedAt: string;
  attendanceId?: string | null;
  checkInAt?: string | null;
  /** Technician field status: checked-in | off-shift | offline */
  status: string;
  isStale: boolean;
}