# lib/api-zod/src/generated/types/upsertTechnicianLocationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * K&S Solar Energy API
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertTechnicianLocationInput {
  latitude: string;
  longitude: string;
  address?: string | null;
}